delete feature added

+++ app/src/main/java/ml/zedlabs/weatherapptask/repository/models/WeatherResponse.kt
@@ -9,7 +9,7 @@ data class WeatherResponse (
     @field:Json(name = "base") var base : String,
     @field:Json(name = "main") var main : Main,
     @field:Json(name = "visibility") var visibility : Int,
-    @field:Json(name = "dt") var dt : Int,
+    @field:Json(name = "dt") var dt : Long,
     @field:Json(name = "timezone") var timezone : Int,
     @field:Json(name = "id") var id : Int,
     @field:Json(name = "name") var name : String,
